+++ src/GUI.java
@@ -1,0 +1,14 @@
+public class GUI
+{
+    Data data = new Data();
+
+    public GUI(Data data)
+    {
+        this.data = data;
+    }
+
+    public void Run()
+    {
+
+    }
+}
